implement new interfaces, factor out RPC, overall structural improvements
