+++ .env.example
@@ -1,14 +1,16 @@
 NEUTRON_RPC="https://rpc-falcron.pion-1.ntrn.tech"
-ETHEREUM_URL="https://sepolia.infura.io/v3/<YOUR_INFURA_ID>"
+ETHEREUM_URL="https://sepolia.infura.io/v3/YOUR_INFURA_API_KEY"
 
+ETHEREUM_DEFAULT_ACCOUNT_ADDRESS="0x51df57D545074bA4b2B04b5f973Efc008A2fde6E"
 NEUTRON_DEFAULT_ACCOUNT_ADDRESS="neutron1m9l358xunhhwds0568za49mzhvuxx9ux8xafx2"
 NEUTRON_PION_1_VAULT_EXAMPLE_CONTRACT_ADDRESS="neutron148w9spa5f9hcwgdy8cnejfel8ly6c2kdazuu94ja5dmy6zyet2ks6c49fd"
 ETHEREUM_SEPOLIA_VAULT_EXAMPLE_CONTRACT_ADDRESS="0x8Fbd2549Dc447d229813ef5139b1aee8a9012eb3"
 
 
 HEIGHT_NEUTRON="26269842"
-HEIGHT_ETHEREUM = "7936546"
 MERKLE_ROOT_NEUTRON="+DBoneb+3cjIEwFve/IEiSMF4ZWj2CVQnKNTK1IE8a4="
 
 # should be globally set for abi verification of Ethereum contracts
-# export ETHERSCAN_API_KEY="Z8KM1ZQMB8UA233P331C3EKX7KBYQSERA4"
+# export ETHERSCAN_API_KEY="YOUR_ETHERSCAN_API_KEY"
+
+ETH_PRIVATE_KEY="YOUR_ETHEREUM_PRIVATE_KEY_FOR_DEPLOYMENTS"
