+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["common", "domains/ethereum"]
+members = ["common", "domains/neutron"]
 resolver = "2"
 
 
@@ -7,7 +7,7 @@ resolver = "2"
 opt-level = 3
 
 [workspace.dependencies]
-ethereum = { path = "domains/ethereum" }
+ethereum = { path = "domains/neutron" }
 base64 = "0.22.1"
 hex = "0.4"
 common = { path = "common" }
