+++ common/src/merkle/types.rs
@@ -1,47 +1,45 @@
-use serde::{Deserialize, Serialize};
-
-/// Represents the output of a Merkle proof verification.
-/// Contains the necessary components to verify a Merkle proof including the root,
-/// key, value, and domain information.
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct MerkleProofOutput {
-    /// The Merkle root hash that commits to the encoded roots
-    pub root: Vec<u8>,
-    /// The key that was queried in the key-value store
-    pub key: Vec<u8>,
-    /// The value hash that commits to the encoded values
-    pub value: Vec<u8>,
-    /// The domain identifier indicating which proving system to use
-    pub domain: Option<u64>,
-}
-
 /// A trait for types that can generate Merkle proofs from RPC calls.
-/// This trait is implemented by different proving systems to fetch proofs
-/// from their respective blockchain networks.
-pub trait MerkleProver {
+///
+/// This trait defines the interface for different proving systems to fetch proofs
+/// from their respective blockchain networks. It provides a standardized way to
+/// retrieve Merkle proofs for any key at a specific block height.
+pub trait MerkleRpcClient {
     #[allow(async_fn_in_trait)]
-    /// Retrieves a Merkle proof for a given key at a specific block height from an RPC endpoint.
+    /// Retrieves a Merkle proof for a given key at a specific block height.
     ///
     /// # Arguments
-    /// * `key` - The key to generate a proof for
-    /// * `address` - The RPC endpoint address
+    /// * `key` - The key to generate a proof for (e.g., account address, storage key)
+    /// * `address` - The RPC endpoint address to fetch the proof from
     /// * `height` - The block height at which to generate the proof
     ///
     /// # Returns
     /// A vector of bytes containing the encoded Merkle proof
-    async fn get_merkle_proof_from_rpc(&self, key: &str, address: &str, height: u64) -> Vec<u8>;
+    ///
+    /// # Note
+    /// The exact format of the proof depends on the implementing blockchain network.
+    /// The proof should be sufficient to verify the existence and value of the key
+    /// in the Merkle tree at the specified block height.
+    async fn get_proof(&self, key: &str, address: &str, height: u64) -> Vec<u8>;
 }
 
 /// A trait for types that can verify Merkle proofs against an expected root.
+///
 /// This trait provides the functionality to verify that a proof is valid
-/// for a given Merkle root.
+/// for a given Merkle root. It is used to ensure that the proof correctly
+/// demonstrates the existence and value of a key in the Merkle tree.
 pub trait MerkleVerifiable {
     /// Verifies the proof against the expected Merkle root.
     ///
     /// # Arguments
-    /// * `expected_root` - The expected Merkle root to verify against
+    /// * `root` - The expected Merkle root to verify against
     ///
     /// # Returns
-    /// A `MerkleProofOutput` containing the verification result
-    fn verify(&self, expected_root: &[u8]) -> MerkleProofOutput;
+    /// A boolean indicating whether the proof is valid for the given root
+    ///
+    /// # Note
+    /// The verification process should check that:
+    /// 1. The proof nodes form a valid path from the leaf to the root
+    /// 2. The leaf node contains the expected key-value pair
+    /// 3. The root hash matches the expected root
+    fn verify(&self, root: &[u8]) -> bool;
 }
