--- domains/ethereum/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "ethereum"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-eth_trie = "0.5.0"
-common.workspace = true
-tokio = { workspace = true, optional = true }
-serde.workspace = true
-serde_json.workspace = true
-reqwest = { workspace = true, optional = true }
-url = { workspace = true, optional = true }
-dotenvy = { workspace = true, optional = true }
-alloy-primitives = "0.8.12"
-alloy-trie = "0.7.9"
-hash-db = "0.15"
-plain_hasher = "0.2"
-triehash = "0.8.4"
-alloy = { version = "0.12.5", features = ["full"], optional = true }
-alloy-sol-types = { version = "0.8.23", optional = true }
-alloy-rlp = { version = "0.3.11", features = [
-    "derive",
-    "arrayvec",
-], optional = true }
-tiny-keccak = "2.0.2"
-# use precompile when running in prod
-tiny-keccak-sp1 = { package = "tiny-keccak", git = "https://github.com/sp1-patches/tiny-keccak", tag = "patch-2.0.2-sp1-4.0.0", features = [
-    "keccak",
-], optional = true }
-
-[features]
-no-zkvm = [
-    "tokio",
-    "reqwest",
-    "url",
-    "dotenvy",
-    "alloy",
-    "alloy-rlp",
-    "alloy-sol-types",
-]
-sp1 = ["tiny-keccak-sp1"]
