--- domains/ethereum/README.md
@@ -1,6 +1,0 @@
-# Ethereum libraries
-
-## Merkle Proofs
-The Ethereum `merkle_lib` library exposes functions to obtain different kinds of storage and account proofs from an Ethereum node, as well as verify them against a `trusted root`.
-
-[click to return home](../../README.md)
