--- domains/ethereum/src/lib.rs
@@ -1,4 +1,0 @@
-//! Ethereum-specific functionality for handling Merkle tree operations.
-//! This module provides utilities for decoding and processing Ethereum's Merkle Patricia Trie leaf nodes.
-
-pub mod merkle_lib;
