--- domains/ethereum/src/merkle_lib/tests/mod.rs
@@ -1,4 +1,0 @@
-mod alloy;
-pub mod defaults;
-pub mod merkle;
-pub mod vault;
