+++ domains/neutron/src/merkle_lib/tests/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "no-zkvm")]
 pub mod defaults;
+#[cfg(feature = "no-zkvm")]
 pub mod merkle;
+#[cfg(feature = "no-zkvm")]
 pub mod vault;
